WOO HOO!!! WatchMe is working and with png compression
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@2618 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ sneer/bricks/src/snapps/watchme/codec/ImageDelta.java
@@ -10,9 +10,9 @@ public class ImageDelta extends Tuple {
 	public final int width;
 	public final int height;
 	
-	public int[]  imageData;
+	public byte[]  imageData;
 	
-	public ImageDelta(int[] imageData_, int x_, int y_, int width_, int height_) { 
+	public ImageDelta(byte[] imageData_, int x_, int y_, int width_, int height_) { 
 		x = x_;
 		y = y_;
 		width = width_;
